refactor: Remove constants package
The constants package of Scanner-Core did not contain a significant amount of constant but rather served as a place for various interfaces.
AnalyzedProperty, AnalyzedPropertyCategory, ProbeType have been moved to the probe package. ScannerDetail has been moved to the config package.
RandomnessConstants has been removed as it should be considered part of the statistical tests of TLS-Scanner which may be ported to Scanner-Core at a later date.